requirements.txt
streamlit==1.2.0
pandas
matplotlib
plotly.express